+++ DelvUI/Config/Profiles/ProfilesManager.cs
@@ -232,6 +232,19 @@ namespace DelvUI.Config.Profiles
             }
         }
 
+        public void CheckUpdateSwitchCurrentProfile(string specifiedProfile)
+        {
+            // find a profile for specified string
+            foreach (Profile profile in Profiles.Values)
+            {
+                // found a valid profile, switch to it
+                if (profile.Name == specifiedProfile)
+                {
+                    SwitchToProfile(profile.Name);
+                }
+            }
+        }
+
         private string? SwitchToProfile(string profile, bool save = true)
         {
             // save if needed before switching